+++ bridges/il2cpp-tracker-tf/CHANGELOG.md
@@ -3,3 +3,4 @@
 ## v1.0.0 (2026-04-18)
 
 Initial public release.
+- Uses Gamelink Lua API (formerly Frida).